% Generated by roxygen2 (4.0.0): do not edit by hand
\name{read.csv.sql}
\alias{read.csv.sql}
\title{sqldf's read.csv.sql (no Tcl/tk!)}
\usage{
read.csv.sql(file, sql = "select * from file", header = TRUE, sep = ",",
  row.names, eol, skip, filter, nrows, field.types, comment.char,
  dbname = tempfile(), drv = "SQLite", ...)
}
\arguments{
\item{file}{A file path or a URL (beginning with \code{http://} or
\code{ftp://}). If the \code{filter} argument is used and no file is to be
input to the filter then \code{file} can be omitted, \code{NULL}, \code{NA}
or \code{""}. The \code{textfile} argument from \code{makeDb} is used here.}

\item{sql}{character string holding an SQL statement. The table representing
the file should be referred to as \code{file}.}

\item{header}{as in \code{read.csv}}

\item{sep}{as in \code{read.csv}}

\item{row.names}{as in \code{read.csv}}

\item{eol}{Character that ends lines}

\item{skip}{Skip indicated number of lines in input file.}

\item{filter}{If specified, this should be a shell/batch command that the input file is piped
through.}

\item{nrows}{Number of rows used to determine column types. It defaults to
50. Using -1 causes it to use all rows for determining column types. This
argument is rarely needed.}

\item{field.types}{A list whose names are the column names and whose contents
are the SQLite types (not the R class names) of the columns. Specifying these
types improves how fast it takes. Unless speed is very important this
argument is not normally used.}

\item{comment.char}{If specified, this character and anything following it on
any line of the input will be ignored.}

\item{dbname}{As in \code{sqldf} except that the default is
\code{tempfile()}. Specifying NULL will put the database in memory which may
improve speed but will limit the size of the database by the available
memory. When using tornado, \code{tempfile()} is not used: \code{dbname} must
be provided as an argument to \code{makeDb}.}

\item{drv}{ignored: the only drive used can be SQLite.}

\item{\dots}{arguments passed to \code{sqldf}.}
}
\value{
If the sql statement is a select statement then a data frame is
returned.
}
\description{
Edited version of \code{read.csv.sql} in the \code{sqldf} package
(\code{tcltk} functionality removed), used mainly to create a SQLite database
from a text file from within R.
}
\details{
Reads the indicated file into an sql database creating the
database if it does not already exist.  Then it applies the sql
statement returning the result as a data frame.  If the database
did not exist prior to this statement it is removed.

Note that it uses facilities of \code{SQLite} to read the file which
are intended for speed and therefore not as flexible as in R.  For
example, it does not recognize quoted fields as special but will
regard the quotes as part of the field. See the ‘sqldf’ help for
more information.

On Windows, if the \code{filter} argument is used and if Rtools is
detected in the registry, then the Rtools bin directory is added to
the search path facilitating use of those tools without explicitly
setting any the path.
}
\references{
http://cran.r-project.org/web/packages/sqldf/sqldf.pdf
}

